finished implementation. problems with docker

+++ src/main/java/com/testtask/controller/UserController.java
@@ -3,8 +3,8 @@ package com.testtask.controller;
 import com.testtask.model.User;
 import com.testtask.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.mongodb.core.query.Criteria;
-import org.springframework.data.mongodb.core.query.Query;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -19,42 +19,54 @@ public class UserController {
     @Autowired
     public UserController(UserRepository repository) {
         this.repository = repository;
-        //example
+        //test example
         repository.save(new User("first", "first"));
         repository.save(new User("second", "second"));
         repository.save(new User("third", "third"));
     }
 
     @GetMapping()
-    public List<User> getAllUsers() {
-        return repository.findAll();
+    public ResponseEntity<?> getAllUsers() {
+        List<User> users = repository.findAll();
+        return new ResponseEntity<>(users, users.size() > 0 ? HttpStatus.OK : HttpStatus.NOT_FOUND);
     }
 
     @GetMapping("/{name}")
-    public Optional<User> getUser(@PathVariable String name) {
-        return repository.findById(name);
+    public ResponseEntity<?> getUser(@PathVariable String name) {
+        Optional<User> user = repository.findById(name);
+        if (user.isPresent())
+            return new ResponseEntity<>(user.get(), HttpStatus.OK);
+        else
+            return new ResponseEntity<>("User doesn't exist", HttpStatus.NOT_FOUND);
     }
 
     @PostMapping()
-    public String addUser(@RequestParam User user) {
+    public ResponseEntity<?> addUser(@RequestBody User user) {
+        if (repository.findById(user.getName()).isPresent()) {
+            return new ResponseEntity<>("User already exist", HttpStatus.CONFLICT);
+        }
         repository.save(user);
-        return user + " added.";
+        return new ResponseEntity<>("User " + user.getName() + " added", HttpStatus.OK);
     }
 
     @DeleteMapping("/{name}")
-    public String deleteUser(@PathVariable String name) {
+    public ResponseEntity<?> deleteUser(@PathVariable String name) {
+        if (repository.findById(name).isEmpty()) {
+            return new ResponseEntity<>("User doesn't exist", HttpStatus.NOT_FOUND);
+        }
         repository.deleteById(name);
-        return "user with name : " + name + " deleted.";
+        return new ResponseEntity<>("User " + name +  " successfully deleted", HttpStatus.OK);
     }
 
-//    @PatchMapping("/{name}")
-//    public Optional<User> updateUser(@PathVariable String name, @RequestParam User user) {
-//
-//        User originalUser = repository.findById(name).get();
-//        originalUser.setEmail(user.getEmail());
-//        originalUser.setName(user.getName());
-//        repository.save()
-//
-//        return repository.findById(name);
-//    }
+    @PutMapping("/{name}")
+    public ResponseEntity<?> updateUser(@PathVariable String name, @RequestBody User user) {
+        Optional<User> userOpt = repository.findById(name);
+        if (userOpt.isPresent()) {
+            User updatedUser = userOpt.get();
+            updatedUser.setEmail(user.getEmail());
+            repository.save(updatedUser);
+            return new ResponseEntity<>("User " + name + " updated", HttpStatus.OK);
+        } else
+            return new ResponseEntity<>("User doesn't exist", HttpStatus.NOT_FOUND);
+    }
 }
